+++ CHANGELOG.md
@@ -7,6 +7,7 @@ TODO Next Version 2.0.0    (??? 2020)
 New Rules
 - #548: New **OptionalField** rule (design) - Do not use an Optional as a field type.
 - #549: New **OptionalMethodParameter** rule (design) - Do not use an Optional as a parameter type for a method or constructor.
+- #550: New **OptionalCollectionReturnType** rule (design) - Do not declare a method return type of `Optional<List>` (or `Collection`, `ArrayList`, `Set`, `Map`, `HashMap`, etc.). Return an empty collection instead. 
 
 Updated/Enhanced Rules and Bug Fixes
  - #467: **VariableName** rule: Change to treat `final` variables the same as regular variables (i.e. not like constants). [BREAKING CHANGE]